Get things mostly working.
Still a problem if you `localias start` and then `localias run`,
seeing a weird EOF error in `cntxt.Search()` probably from
the ReadPidFile
